Рефракторинг и слияние библиотек в core
