+++ core/__init__.py
@@ -1,0 +1,5 @@
+from . import core
+
+__version__ = "in developing ..."
+
+__all__ = ["core"]
