+++ core/core.pyi
@@ -1,7 +1,8 @@
 """ Type hintings """
 
 
-from tkinter import Event, Canvas, Tk
+from tkinter import Event, Canvas, Tk, StringVar
+from tkinter.ttk import Notebook
 from typing import (
     Callable,
     Tuple,
@@ -10,6 +11,7 @@ from typing import (
 
 
 # _defaults.py
+" ~ graphic_lib ~ "
 DEFAULT_USED_EVENTS: Tuple[str]
 DEFAULT_FIRST_COLOR: str
 DEFAULT_SECOND_COLOR: str
@@ -21,23 +23,50 @@ DEFAULT_MOUSE_SPEED: int
 DEFAULT_CANVAS_W: int
 DEFAULT_CANVAS_H: int
 DEFAULT_CANVAS_BG: str
+" ~ image_lib ~"
+DEFAULT_FILTERS_1: Iterable[str]
+DEFAULT_FILTERS_2: Iterable[str]
+DEFAULT_FILTERS_3: Iterable[str]
 
 
 # _custom_objects.py
 class CustomCanvas(Canvas):
-    old_point: (None, Tuple[int, int])
-    obj_oval: (None, int)
-    obj_line: (None, int)
-    obj_rectangle: (None, int)
-    line_sequences: list
-    obj_storage: dict = {}
-    modified_objs: list
-    start_point: (None, Tuple[int, int])
-    obj_tag: (None, str)
-    hover: bool
+    def __init__(self, *args, **kwargs) -> None:
+        super().__init__(*args, **kwargs)
+
+        " ~ graphic_lib ~ "
+        self.old_point: (None, Tuple[int, int])
+        self.obj_oval: (None, int)
+        self.obj_line: (None, int)
+        self.obj_rectangle: (None, int)
+        self.line_sequences: list
+        self.obj_storage: dict
+        self.modified_objs: list
+        self.start_point: (None, Tuple[int, int])
+        self.obj_tag: (None, str)
+        self.hover: bool
+        " ~ image_lib ~ "
+        self.img: dict
+        ...
+
+class CustomNotebook(Notebook):
+    def __init__(self, root: Tk, **kwargs) -> None:
+        super().__init__(**kwargs)
+        self._count: int
+        self.root: Tk
+        self.image_processing: Img
+        self.events: Events
+        self._canvases: list
+        ...
+
+    def create_new_canvas(self) -> None: ...
+    def select_curr_tab(self, event: Event) -> None: ...
 
 
-# _basic.py
+""" ============= graphic_lib ============ """
+
+
+# graphic_lib/_basic.py
 def reset(function: Callable[..., None]) -> Callable[..., None]: ...
 def transform_coords(old_coords: Tuple[int, int], new_coords: Tuple[int, int]) -> Tuple[int, float]: ...
 def transform_line_coords(old_coords: Tuple[int, int], new_coords: Tuple[int, int]) -> Tuple[int, float]: ...
@@ -45,7 +74,7 @@ def transform_line_sequence(point_storage: Iterable) -> Tuple[(int, float), (int
 def detect_object(event: Event, canvas: CustomCanvas) -> [str, None]: ...
 
 
-# _draw.py
+# graphic_lib/_draw.py
 class Draw:
     def __init__(self, event: Event, canvas: CustomCanvas) -> None: ...
     def point(self, *, size: (int, float) = DEFAULT_BRUSH_SIZE, color: str = DEFAULT_FIRST_COLOR, debug_mode: bool = False) -> None: ...
@@ -55,12 +84,14 @@ class Draw:
     def polygon(self, *, thickness: (int, float) = DEFAULT_THICKNESS, bgcolor: str = DEFAULT_SECOND_COLOR, outcolor: str = DEFAULT_FIRST_COLOR) -> None: ...
     def move(self, *, mouse_speed: int =  DEFAULT_MOUSE_SPEED) -> None: ...
     def fill_objects(self, *, color: str = DEFAULT_CHANGE_COLOR) -> None: ...
+    def thickness_objects(self, *, thickness: (int, float) = DEFAULT_THICKNESS) -> None: ...
+    def outline_color_objects(self, *, color: str = DEFAULT_CHANGE_COLOR) -> None: ...
     def quick_eraser(self) -> None: ...
     def on_canvas(self) -> None: ...
     def undo(self) -> None: ...
 
 
-# _events.py
+# graphic_lib/_events.py
 class Events:
     def __init__(self, root: Tk, used_events: Tuple[str], canvas: CustomCanvas) -> None: ...
     def event_btnClear(self) -> None: ...
@@ -72,5 +103,32 @@ class Events:
     def event_undo(self) -> None: ...
     def event_move(self, *, mouse_speed: int = DEFAULT_MOUSE_SPEED) -> None: ...
     def event_btnFill(self, *, color: str = DEFAULT_CHANGE_COLOR) -> None: ...
+    def event_btnThickness(self, *, thickness: (int, float) = DEFAULT_THICKNESS) -> None: ...
+    def event_btnOutlineColor(self, *, color: str = DEFAULT_CHANGE_COLOR) -> None: ...
     def event_btnQuickEraser(self) -> None: ...
     def event_onCanvas(self) -> None: ...
+
+
+""" ============= image_lib ============ """
+
+
+# image_lib/_image_processing.py
+drawQ: bool
+
+def random_color() -> str: ...
+
+class Img:
+    def __init__(self, canvas: CustomCanvas) -> None: ...
+    def set_image(self) -> None: ...
+    def get_info(self) -> None: ...
+    def save_image(self) -> None: ...
+    def return_image(self) -> None: ...
+    def set_bg(self) -> None: ...
+    def grab(self, event: Event) -> None: ...
+    def drag(self, event: Event) -> None: ...
+    def zoom(self, event: Event) -> None: ...
+    def redraw(self, direction: str = "in") -> None: ...
+    def apply_filter_1(self, f: StringVar) -> None: ...
+    def apply_filter_2(self, f: StringVar, per: StringVar) -> None: ...
+    def apply_filter_3(self, f: StringVar) -> None: ...
+    def apply_filter_4(self) -> None: ...
