--- image_lib/_image_core.py
@@ -1,3 +1,0 @@
-from ._image_processing import *
-from ._custom_objects import *
-from ._defaults import *
